move files to classbased folder
