+++ src/index.js
@@ -1,10 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom";
 // component file
-import TodoContainer from "./components/TodoContainer";
+import TodoContainer from "./classBased/components/TodoContainer";
 
 //stylesheet
-import "./App.css";
+import "./classBased/App.css";
 
 ReactDOM.render(
   <React.StrictMode>
